requirements.txt
# Core
PyQt5>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0

# Dev / Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0

# Linting / Formatting
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Packaging
pyinstaller>=6.0.0
